pyproject.toml
[project]
name = "techsnack-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "streamlit>=1.40.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.11.0",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "newsapi-python>=0.2.7",
    "feedparser>=6.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

